feat: Refactor hardcoded values to constants
This commit begins the "AI Slop Reduction" initiative by refactoring hardcoded values into centralized constants, improving code quality and maintainability.

- Establishes a new pattern for constants:
  - `lib/constants/index.ts` for shared UI constants.
  - `hooks/useBluetoothHRM.constants.ts` for hook-specific values.
  - `utils/constants.ts` for server-side constants.
- Moves the `constants` directory to `lib/constants` to ensure it's included in the server-side build.
- Refactors `useBluetoothHRM`, `HeartRateZones`, `HRMonitorStatusIndicator`, `socketManager`, `websocketUtils`, and other components to use the new constants.
- Preserves environment variable configuration for WebSocket grace period.

fix: Correct import paths for build compatibility

- Resolves `ERR_UNSUPPORTED_DIR_IMPORT` and `ERR_MODULE_NOT_FOUND` in the `tsc` and Webpack builds by ensuring import paths are compatible with both server (tsc) and client (Webpack) build systems.
- Fixes linting errors and build failures caused by incorrect import paths.
