+++ components/HrmTiles.tsx
@@ -2,7 +2,7 @@
 'use client'
 import HrTile from '@/components/HrTile'
 import { useWebSocket } from '@/context/WebSocketContext'
-import { MAX_HR_DEFAULT } from '@/constants/index'
+import { MAX_HR_DEFAULT } from '@/lib/constants/index'
 import { getHrZoneProps } from '@/utils/visualization'
 import Grid from '@mui/material/Grid'
 import Skeleton from '@mui/material/Skeleton'
