+++ lib/constants/bluetooth-messages.ts
