+++ lib/constants/timer.ts
