+++ package.json
@@ -20,8 +20,8 @@
     "pm2:stop": "pm2 stop hrm-server",
     "pm2:restart": "pm2 restart hrm-server",
     "pm2:delete": "pm2 stop hrm-server && pm2 delete hrm-server",
-    "lint": "eslint app/ components/ constants/ context/ hooks/ lib/ services/ tests/ types/ utils/ server.ts proxy.ts --cache",
-    "lint:fix": "eslint app/ components/ constants/ context/ hooks/ lib/ services/ tests/ types/ utils/ server.ts proxy.ts --fix --cache",
+    "lint": "eslint app/ components/ context/ hooks/ lib/ services/ tests/ types/ utils/ server.ts proxy.ts --cache",
+    "lint:fix": "eslint app/ components/ context/ hooks/ lib/ services/ tests/ types/ utils/ server.ts proxy.ts --fix --cache",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
     "test": "pnpm run test:all",
